Do not show duplicates.
When running slide show only pictures that have not been seen
before should be shown. So the application now keeps track
of pictures already shown and skips any that it has already
displayed. This record is discarded upon reaching the end of
the current slide show and before repating the slide show.